Divide by zero bugfix.
With a lot of classes, there are so few samples that
sniffer parser ignored first 100 and last 100 values,
which ignored basically the entire set of samples.

So the list was empty which caused a divide-by-zero
when computing average.  Instead of ignoring 100 values
the code now ignores the first 10% and last 10% of
samples (configurable via ignore_frac).